IDE: codehelp: fixed type
git-svn-id: http://svn.freepascal.org/svn/lazarus/trunk@29802 4005530d-fff6-0310-9dd1-cebe43e6787f

+++ ide/codehelp.pas
@@ -138,7 +138,7 @@ type
     FPDocFilename: string;
     FPDocFileOwner: TObject;
     FPDocFilenameTimeStamp: integer;
-    FilesTimeStamp: integer;
+    FilesTimeStamp: int64;
   end;
   
   { TCodeHelpElement - mapping between one codetools position and a fpdoc xml node.
